复重任务 列表 分页 添加搜索

+++ app/src/main/java/qx/app/freight/qxappfreight/bean/response/WaybillsBean.java
@@ -120,7 +120,7 @@ public class WaybillsBean implements Serializable {
     /**
      * 预期费用
      */
-    private int amountOfMoney;
+    private double amountOfMoney;
     /**
      * 逾期费用
      */
